Carbon Black Output - Test error update
